Update gameoflife.py
Introduce an enhanced graphical user interface with the pygame_gui library, featuring a reset button for improved user interaction. Also incorporated scipy and pytorch for a more sophisticated update() function and better generations of "life".The update provides a visually appealing interface and additional functionality for a better user experience.